requirements.txt
aiogram==2.25.1
aiohttp==3.8.3
aiohttp-proxy==0.1.2
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
autocommand==2.2.2
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==2.1.1
cheroot==9.0.0
CherryPy==18.8.0
decorator==5.1.1
dnspython==2.3.0
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.0
frozenlist==1.3.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hypercorn==0.14.3
hyperframe==6.0.1
idna==3.4
inflect==6.0.2
iniconfig==2.0.0
ipaddress==1.0.23
jaraco.classes==3.2.3
jaraco.collections==3.8.0
jaraco.context==4.3.0
jaraco.functools==3.5.2
jaraco.text==3.11.1
loguru==0.6.0
magic-filter==1.0.9
more-itertools==9.0.0
multidict==6.0.4
openai==0.26.5
openpyxl==3.1.0
packaging==23.0
peewee==3.15.4
phonenumbers==8.13.5
pluggy==1.0.0
portend==3.1.0
priority==2.0.0
psycopg2-binary==2.9.5
pydantic==1.10.4
pyQiwiP2P==2.0.6
PySocks==1.7.1
pytest==7.2.1
python-dotenv==0.21.1
python-socks==2.1.1
pytz==2022.7.1
requests==2.28.2
response-report==1.1.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
tempora==5.2.1
toml==0.10.2
tomli==2.0.1
tqdm==4.64.1
typing==3.7.4.3
typing_extensions==4.4.0
urllib3==1.26.14
validators==0.20.0
wsproto==1.2.0
yarl==1.8.2
zc.lockfile==2.0
